+++ src/library/general/dlhook.cpp
@@ -166,8 +166,13 @@ __attribute__((noipa)) void *dlopen(const char *file, int mode) __THROW {
         return nullptr;
     }
 
-    LOGTRACE(LCF_HOOK, "%s call with file %s", __func__, (file!=nullptr)?file:"<NULL>");
+    LOGTRACE(LCF_HOOK, "%s call with file %s and flags 0x%x", __func__, (file!=nullptr)?file:"<NULL>", mode);
     void *result = nullptr;
+
+    if (mode & RTLD_DEEPBIND) {
+        LOG(LL_DEBUG, LCF_HOOK, "   removing RTLD_DEEPBIND flag");
+        mode &= ~RTLD_DEEPBIND;
+    }
 
 #if defined(__APPLE__) && defined(__MACH__)
     if (dlopen_from) {